+++ ChangeLog.md
@@ -1,5 +1,6 @@
 2018-10-09
 
+- add asset-class-select-button
 - executor associated with assetClass (not direct asset instance)
 
 2018-10-08